fix for 32-bit targets (tested on cfarm26)

+++ src/binary64/acos/acos.c
@@ -282,7 +282,7 @@ double as_acos_refine(double x, double phi){
   e = 52-(107+e);
   e = e<0?0:e;
   e = e>52?52:e;
-  u64 m = (1ul<<52)-(1ul<<e);
+  u64 m = ((u64)1<<52)-(1ul<<e);
   if(__builtin_expect(!((t.u+(1l<<(e-1)))&m), 0)){
     if(x==-0x1.771164bfd1f84p-3 ) return 0x1.c14601daaf657p+0  - 0x1p-54;
     if(x==-0x1.4510ee8eb4e67p-1 ) return 0x1.211c0e2c2559ep+1  - 0x1p-53;
